Restructure, split libc, arch dir
The kernel and libc are separated. Libc will build an ar file that will
be linked against the kernel. A library called libk will be generated
which is basically libc for the kernel. It ties directly into the kernel
whereas libc would use system calls.

Also moved the arch dependent code into a subdirectory.