Install syscall ISR, whoami, shutdown, empty line
Interrupt 0x42 (66) is setup as the syscall handler.
Double meaning, 42 is the answer to life, the universe,
and everything, and 66 is the order given in Star Wars.

Added shell commands "whoami", "shutdown", and "sc".

Whoami always prints "groot". Shutdown only works in
an emulator (specifically Bochs or QEMU). sc just raises
a bogus syscall. No syscalls are implemented yet.