Stripped stage 1 to only loading stage 2
The stage 1 bootloader will only relocate itself,
load the stage 2 bootloader and jump to it.

The stage 2 boot loader is basically the same as the original
bootloader. There's a macro definition to control the start
sector and kernel sector size. At the moment that are manually
changed but will eventually change at runtime by reading partition
tables and filesystem directories.

Still need to make a "stage 3" loader that would live in the VBR.
It must have holes for BIOS parameters table and other things.
Right now it's not vital.